*** empty log message ***

+++ html/menu_mail.php
@@ -1,4 +1,4 @@
-<!-- start of $Id: menu_mail.php,v 1.29 2002/05/20 14:37:33 rossigee Exp $ -->
+<!-- start of $Id: menu_mail.php,v 1.30 2002/06/27 22:17:52 rossigee Exp $ -->
 <table border="0" align="center" cellpadding="0" cellspacing="0" width="100%">
     <tr>
         <td bgcolor="<?php echo $glob_theme->inside_color ?>">
@@ -20,7 +20,10 @@
                         <a href="<?php echo $_SERVER['PHP_SELF'] ?>?action=forward&amp;mail=<?php echo $content['msgnum'] ?>" class="menu"><?php echo $html_forward ?></a>
                     </td>
                     <td class="menu" align="center" bgcolor="<?php echo $glob_theme->menu_color ?>">
-                        <a href="delete.php?delete_mode=1&mail=<?php echo $content['msgnum'] ?>&amp;only_one=1" class="menu"><?php echo $html_delete ?></a>
+                        <a href="down_mail.php?mail=<?php echo $content['msgnum'] ?>&<?php echo session_name() . "=" . session_id(); ?>" class="menu" target="_blank"><?php echo $html_down_mail ?></a>
+                    </td>
+                    <td class="menu" align="center" bgcolor="<?php echo $glob_theme->menu_color ?>">
+                        <a href="delete.php?delete_mode=1&mail=<?php echo $content['msgnum'] ?>&amp;only_one=1" class="menu" onClick="if (confirm('<?php echo $html_del_msg ?>')) return true; else return false;"><?php echo $html_delete ?></a>
                     </td>
                     <?php if ($conf->enable_logout) { ?>
                     <td class="menu" align="center" width="80" bgcolor="<?php echo $glob_theme->menu_color ?>">
@@ -32,4 +35,4 @@
         </td>
     </tr>
 </table>
-<!-- end of $Id: menu_mail.php,v 1.29 2002/05/20 14:37:33 rossigee Exp $ -->
+<!-- end of $Id: menu_mail.php,v 1.30 2002/06/27 22:17:52 rossigee Exp $ -->
